requirements.txt
fastapi[all]~=0.109.2
pydantic~=2.6.1
sqlalchemy~=2.0.26
psycopg2
alembic~=1.13.1
python-dotenv~=1.0.1
fastapi-sqlalchemy
PyJWT~=2.8.0
passlib~=1.7.4